Todos now persisted to localStorage
Used useRef hook to keep track of whether page is loaded for the first time.
If so, initialize a local object [] via storage.setItem, instead of saving the current state todos to localStorage as it would be empty for the next refresh.
If not, simply retrieves the locally stored todos object and set its corresponding state to itself.